feat: unify ad media url building
